requirements.txt
cycler==0.10.0
matplotlib==2.0.2
numpy==1.13.3
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.0
scipy==0.19.1
six==1.10.0
pytest==3.4.0
Django==2.0
South==0.7.5
django-crispy-forms==1.1.3
django-debug-toolbar==0.9.4
django-extensions==0.8
django-floppyforms==0.4.7
django-registration==0.8.0
gunicorn
django-jquery==3.1.0
filestack-python==2.3.1
django-typeform==0.1.0
django-tinymce4==3.0.2
django-front
pytest==3.4.0
pytest-django==2.8.0
psycopg2
django-tinymce4-lite